Merge initial work on supporting HTTP calls between services (#745)
This PR merges the initial work we've been doing on supporting making
HTTP calls between services; it is currently behind a feature flag and
will not impact existing applications.

There's a lot of new stuff in this PR, so I'll try and summarise the
changes here:
- If `ENCORE_DAEMON_DEV` is set, we will now watch the `runtime` source
code during `encore run`. This will allow quicker iteration on the
runtime library.
- The service discovery config map is now populated for local
development, for now always pointing at the daemons proxy server.
- A new experiment, `external-calls` has been introduced; when switched
on all service-to-service calls will be routed over HTTP, rather than
internally within the running binary. This is experimental and really is
not ready for use by _anybody_.
- A new `CallMeta` data structure exists within the API SDK, which
tracks external information about a call, as well as internal meta data.
This structure exists to act as a layer between the transport (HTTP in
this case) and the handler invocation code. This object is constructed
at the highest level `handler` in the HTTP service, and passed via the
context to the running handler invocation function.
- A new `errmarshalling` package has been added. The role of this
package is to enable an "internal" version of marshalling to occur for
the wire level, which can be used on service-to-service calls to
propagate more information between services. For calls not originating
from another service in the same Encore app, this package is not used.
- A new `svcauth` package has been added to the API SDK; it's use is to
allow services to sign and verify service-to-service requests. Currently
there is just a no-op implementation (which always approves the auth),
but we plan to add several authentication options in the future _before_
HTTP based service-to-service calls to live.
- A new `transport` package has been added to the API SDK; it's designed
to abstract away the transport layer from the `svcauth` and server code,
allowing us to quickly and easily support other transport methods (such
as GRPC) in the future. It's make use case right now is to allow us to
read and write meta data on requests being made in that transport.
- The Github actions have been updated, to also run the unit tests for
the `runtime`

---------

Co-authored-by: André Eriksson <andre@encore.dev>