Ignore `traceparent` header if not internal service call
We've encounted an issue where CloudRun is adding `traceparent` to inbound calls, this is causing an issue with our trace engine which then can't find the parent span.

For now we're going to work around this by simply disabling reading the `traceparent` when it's not coming from an internal service-to-service call.

In the future we want to support this and allow the intergration of other trace systems, but this will fix the issue for right now.