Introduce no-op infrastructure (#775)
This commit introduces the concept of no-op infrastructure. This is
required so that Encore can build a single binary containing all the
services, however deploy a container with just a single service running
inside it - all infrastructure used by other services which are not
running will not work as the container will not be provisioned
permissions to that infrastructure. Instead any operations on that
infrastructure should result in an error.