Run 1 process per service under experimental flag
This commit updates Encore so that under the `external-calls` experiment, each service will be spun up as it's own process locally with a service map that the services can use to call each other.

Each instance of the built binary is only provided the minimal runtime configuration that it needs; that is if a service doesn't use a PubSub topic, then it won't be given any configuration for that topic - which will return in an error _if_ the topic is called in any way.