Remote Auth Handler calls implemented
This commit adds the ability for a Encore service (or gateway) to make a remote call to another container to run an auth handler.

This is required when the auth handler requires access to infrastructure, that only the service it is defined in has access to. So all other containers will route their auth calls to that container.