Remote Auth Handler calls implemented (#786)
This PR adds the ability for an Encore service (or gateway) to make a
remote call to another container to run an auth handler.

This is required when the auth handler requires access to infrastructure
that only the service it is defined in has access to. So all other
containers will route their auth calls to that container.

This PR also fixes an issue where secrets would cause a panic in
services when not set, even if those services where not running within
the container.