Don't rewrite Host in Gateway proxy
Previously we where rewriting the HTTP `Host` header in the proxy to preserve the original host information - which would allow Encore to perform Host based switching behaviour in downstream services.

However this breaks in CloudRun as the load balancers will then end up rerouting the request back to the proxy - causing an infinite recursion of calls.