Support different outbound SvcAuth methods
This commit updates the runtime to allow outbound requests to be signed with different auth methods than inbound requests can use.

This is useful in a Kubernetes environment where the API Gateway might use `encore-auth`, but services inside the cluster might use `noop`. In this case the Gateway has to only accept `encore-auth`, but it needs to also load `noop` during system startup to sign the outbound requests it makes.