cli: bump 'encore gen wrappers timeout'
We've gotten reports that the command times out.
It's not clear whether it's due to taking a long time to parse
larger projects with a cold cache, or due to a deadlock.

I haven't been able to reproduce the deadlock, so let's start
by bumping the timeout and seeing if that helps.