miscellaneous fixes (#852)
## svcproxy: fix mutex handling
The svcproxy held a read lock for the duration of serving a request,
which was problematic when doing live reloads with long-running requests
like WebSockets.

Fix this by resolving the proxy with the mutex and then
forwarding the request without holding the lock.

## redis: fix panic
We were calling `.Addr()` on the miniredis server, but it turns out that
panics
if calling it after `.Close()` has been called. Fix this by caching the
address in
`Start` and looking it up afterwards.