run: wait for service to start before switching proc
Now that the binary itself is invoking `net.Listen` we
can end up in a race during live reload, where we switch
the active process over but the process isn't listening yet.
This can lead to requests failing in that window.

Fix this by polling for the port to be open before
considering the startup complete.