Open the dev dash on 'encore run' startup
To make the dev dashboard more discoverable
for new users, and to make it more convenient
for more experienced Encore users, open the
dev dashboard on `encore run` startup.

This behavior can be controlled with the new
`--browser` flag to `encore run`. To revert to
the previous behavior, use `--browser=never`.