ENC-1735 Add graphql proxy to local dev dash (#889)
To support the new Stored Requests feature, we need to start storing api
requests on the platform. This PR exposes access to the platform graphql
endpoint through the local dev dash /__graphql. We downgrade the access
level for these graphql requests to a limited set currently only
consisting of reading and storing stored requests.