Support graceful pod termination in Kubernetes
This commit modifies the graceful shutdown system of the Encore runtime
such that if we're running inside a Kubernetes pod, we will wait for a
period of time after receiving a SIGTERM before we start the shutdown
of the running application.

This allows time for Kubernetes to put the pod into a terminating state
while external systems such as load balancers get updated that the pod
is not in a health state and should not be sent new traffic.

Effectively, this means that the pod will be in a terminating state for
the duration of the `graceTerminationPeriodSeconds` value in the pod spec
(minus the gracefult shutdown time configured by Encore).