metrics: fix initialization race
When using a counter group or gauge group, the initialization of the
time series happens on first use. However, if two different goroutines
tried to access the same set of labels concurrently, there was a rare race
where one goroutine could try to increment the time series state before
the other goroutine had finished initializing the data structure.

Fix this by waiting for the initialization to complete before returning
the counter on a call to .With(). This wait is cheap and involves only
a single atomic load in the fast path after initialization has completed.

Thanks Dan Upton and Neal Lathia for the report.

Fixes #977