[ENC-2019] Cache metadata for apps.
This commit modifies the daemon to cache metadata for apps on the filesystem,
this then allows the local dev dash to access this metadata without having to
have the app running via `encore run`.

This commit then tidies up a bunch of code around the dev dash API to
standardise the status response allowing us to make much faster updates
in the dev dash relating to the status of the app.

Finnally I've added the ability for the dev dash to pass a correlation id
to the daemon when it makes an api-call, allowing the dev dash to filter
for logs relating to the request it made.