Remove rest api tutorial for typescript (#1062)
Because we recently made changes automativity this tutorial (and not the
Go version) is what is being shown on our docs pages. Removing it fixes
this problem.

This content is not lost, I have another PR coming soon that has both
the Go and TS versions on the same page.