cli: tweak test database setup
Previously we created the database as `foo_template` and then
created the database with the intended name as `foo` by cloning
the template. This mirrors how we do it when creating temporary dbs.

However, this led to an issue where if a migration uses the fully-qualified
name of the database the migration would fail to execute (as the database
is now named differently).

Fix this by instead creating it under the original name and then
renaming it to `foo_template`.