runtimes/core: handle worker mode
When the runtime starts up it by default begins listening to
HTTP requests and so on. But when running in a worker thread
this is undesirable. To support this use case, tell the runtime
on startup whether it's running in a worker or not and disable
listening for ports if that's the case.