[errs] use errors.As instead of type assertions (#1808)
I'd like to be able to embed the `errs.Error` struct without losing the
error metadata when the handlers in `errs` are used. Currently, if I try
to use an error struct with an embedded `*errs.Error` struct, the type
assertions fail and the error metadata is lost.

The use case is a bit niche, but if this change is controversial I can
go into more details.