fix/make RPC functions and parameters accessible in generated Go client (#1861)
# Changes
This PR addresses a problem with Go client generation from a Typescript
API. Currently naming of RPC functions and parameters are taken straight
from the Typescript codebase. This is a problem because in Typescript
these typically start with lowercase characters, making them private and
unaccessible in Go.

This PR fixes that by making sure that the first letter is always
capitalized for:
- RPC function names
- RPC parameters

# Note for potential further improvement of client generation (Go ->
Typescript)
Ideally client generation the other way around (from Go code to
Typescript) would also follow Typescript naming conventions of using
`camelCase`. Currently it just takes the casing from Go, which means it
usually is `PascalCase`.

I would be happy to submit a PR for this, but also realize this could
have backwards compatibility issues so wanted to get your thoughts on
the matter first.