feat(daemon): clarify warning when secrets are missing due to lack of app access
Sometimes I see other co-workers struggling with this error, and it can
be very frustrating if you have no idea about the origin of the problem
which, in some cases, is simply because you're not currently logged into
your Encore account with access to the app you're trying to run.

I'd like to suggest to the user to run `encore auth whoami`, but I've
tried not to make the help message longer than it already is. With some
luck, this could help other people find guidance when encountering this
error.