Add e2e tests for js apps (#2084)
Since we need to build the jsruntime for this to work, I moved the e2e
tests to its own job so that it runs in parallel. That job takes about
the same time as the other test jobs, so hopefully it shouldnt add much
time to ci.

This also fixes error cause handling, it was duplicated