runtimes/go: map context errors from the sqldb
Currently all errors from the db are by defauled mapped to the
"Unavailable" encore error. This includes context errors.

For "canceled" in particular, I don't think unavailable is the
correct error code to return because it isn't indicative of an
error with the system.

So this change maps the context errors to equivalent encore error
codes.