Add encore.dev/types/option package
This adds a new `encore.dev/types/option` package for Encore.go,
with a new type option.Option[T] for representing optional values.

This improves Encore's support for representing optional types in APIs.
It's also something that we've internally found to be useful again and again
in large code bases, where it's difficult to keep track of whether a pointer value
*T means the type is nullable or not.